refactor(store): deepen StoreTenancy — single tenancy seam via CanonicalPath
StoreTenancy now owns loadConfig, hash8, sanitizedBasename,
canonicalWsSync via CanonicalPath seam, sidecar, migration and
tenancyFor(cwd) → {dir, mode, runtimeDir, canonical}. File-per-
workspace central kept, no PID level. paths.ts is thin adapter
re-exporting StoreTenancy + CanonicalPath (configDir/hashStorePath
byte-identical). hash-store.ts now imports from StoreTenancy,
removing direct loadConfig leakage. CanonicalPath owns sync/async
resParts (locality: ELOOP fix in one place, two adapters justify seam).

Refs: docs/specs/issue-24-store-tenancy.md, /tmp/architecture-review-1787565280.html C1
